requirements.txt
beautifulsoup4==4.6.3
bs4==0.0.1
python-dotenv==0.9.1
selenium==3.141.0
urllib3==1.24.1
